refactor(wpf-migration phase 0b): extract integrity/manifest/fingerprint/disk-preflight to Core
Moves the backup-integrity logic into Savedrake.Core (which now references
DotNetZip 1.16.0 + Newtonsoft.Json 13.0.4, pinned to the app's versions). The
WinForms app keeps thin forwarders and still builds/runs. No behavior change.

- Manifest: ManifestEntryName, IsManifestEntry, BuildBackupManifest,
  StableManifestHash, VerifyZipRestorable, VerifyZipAgainstManifest, HasManifest,
  RestoreBlockedByManifest, ClassifyBackupFully (+ private Sha256Hex).
- Fingerprint: ComputeSaveFingerprint (-> SaveScan/Manifest internally).
- DiskPreflight: SafetyMarginBytes, GetDirectorySize, GetZipUncompressedSize,
  HasFreeSpaceFor.
- Main.cs keeps private forwarders/const aliases; harness repoints these lookups
  to the Core types. Restore engine (ExtractZipToStaging etc.) stays on Main.

Verified: Debug + Release build clean (expected NU1903 DotNetZip advisory only);
harness 196/196 (these methods now run against Core, exercising corrupt/missing/
tampered detection, fingerprint stability/fail-closed, classification); the app
launches and runs cleanly (forwarders wired).
